Changes: MWidgetController and MWidgetView - apply the style size hint in the controller
RevBy: Ville Voutilainen, Daniel d'Andrada
Details: Previously, each individual widget's view's sizeHint() function was responsible for reading
the style's sizeHint function.  This led to a lot of inconstancy between the classes.
This change moves the logic to MWidgetController so that all widgets treat the CSS specified sizes
in the same way.