Changes: limit rendering of minimized applications to 5fps
RevBy: Stanislav, Michael Hasselmann

Changes: Minimized applications are not allowed to repaint more than
5 times per second. All paint events when beeing in the switcher
and invisible will be discarded.

The base is there to disable paiting completely also for invisible
foreground applications. It is not yet actived though as it would
eat the very first frames when an application is started.