New: Ctrl+Q closes all windows
RevBy: Daniel d'Andrada
Details: When CTRL+Q is pressed, all windows are closed which leads to closing application.
MApplication::quit() is not used, because that wouldn't give an application a chance to
ask user if he want to save his data or not (as it is in other desktop environments like gnome or
windows).