Fixes: NB#182145 - MLabel does not display long rich text properly
RevBy: Sergiy Dubovik, Armin Berres

Details: If the size of the rich text exceeds the maximum size for pixmaps of 2048x2048 pixels, the rich text will only partly be shown. The patch fixes this issue by using up to two pixmap-tiles, where each tile has the height of the display. If the rich text spans over more than two tiles, the currently invisible tile will be updated with the currently visible text. The update of the tile can be done in a quite short time, so that the jittering is very minimal. If there is no cache available at all for not even two tiles, the text will be drawn directly as fallback.