Changes: package the examples into an own debian package
RevBy: Armin, John

Details: Jani Mikkonen requested to package the parts of the
chaining example in libmeegotouch, so that he can then use
these programs for some tests where he is using fakemail
and fakegallery right now. he wanted to get rid of that
duplicated code.

when i looked a bit deeper into the build system i found
that the examples could not be built and not be packaged
in the current version. so i added that possibility to
the build system. some other examples do not have install
rules in their .pro files yet, and perhaps the
debian/meegotouch-examples.install would then need to be
expanded for the other examples too.

i verified that the chaining example does work, when installed
from the created .deb package.

please test and comment.