Fixes: NB#185011 - Widget placed in view menu via MWidgetAction doesn't update its style on theme change.
ModifiedBy: Daniel d'Andrada
RevBy: Michal Guminiak, Daniel d'Andrada
Details:

When the parent of given item changes, we must reload styles for
child items, because otherwise their style wouldn't be updated correctly.

We sometimes style widgets in .css files by specifying grandparents or
great-grandparents for widgets, like when styling MLabel in MApplicationMenu.
MLabel is not direct child of MApplicationMenu, it is child of MListItem
or MComboBox which are children of MApplicationMenu. In such case when we
reparent MComboBox, its MLabel isn't aware that it is now a grandchild
of MApplication menu so it has wrong style. We must tell it explicitly to
reload style.