Changes: force connection to remote themedaemon on arm
RevBy: Holger

Details:
Recently due to bugs some apps did not connect to the themedaemon.
These bugs have just been noticed by accident. As all and every
app on the device should connect to the themedaemon we force this
connection on arm.
If for whatever reason an application is supposed to be started
without connecting to the themedaemon the parameter "-local-theme"
can be used.
Additionally some minor cleanup has been done.