Fixes: NB#190714 - ApplicationPages fails to appear if appear()s are called many times fast enough
ModifiedBy: Daniel d'Andrada (essentially added the unit test)
RevBy: Daniel d'Andrada
Details:
When a application page was in the appearing state and another one in the
disappearing state consecutive calls to appear() have been ignored for
the window with the appearing state but not the one in the disappearing
state. Now if there are pending transitions a transition for for currently
appearing page will be queued.