Changes: make handling of "@" and ";" characters work, add unit test
RevBy: Mike Fabian
Details: This patch builds on Janne Heikkilä's patch in merge
request 500 and extends it a bit.
The former patch was missing some special handling for the
@ and ; characters, for example we could not get from
"fr_FR@collation=phonebook" to "fr_FR". it was directly skipping
to "fr". This patch fixes it and also adds a unit test for the
function in MImageDirectory