Changes: remove plain qt style in libmeegotouch
RevBy: Tomas Junnonen, Berthold Krevert
Details: the plain qt style can now be found at:
http://meego.gitorious.org/meegotouch/meegotouch-qt-style

The reason for the split of the plain qt style from libmeegotouch
is the following problem: The style should display the content
picker instead of the normal qt file dialog in qt applications.

The libcontentmanager does implement the content picker. And
libcontentmanager links against libmeegotouch. So we would have
an impossible cyclic dependency when we would try to link against
the libcontentmanager from the plain qt style in libmeegotouch.

So we had to split the plain qt style form the rest of the
libmeegotouch and put it into its own repo at the URL above.

Now the plain qt style can link against both libmeegotouch and
libcontentmanager.