Changes: Add a CSS validator tool that verifies CSS the files' syntax.
RevBy: TrustMe

Details:
These changes add the MStyleSheetParser::validate method checks the
source of CSS files by performing the save operations as
MStyleSheetParser::parse but ignoring the contants' replacement and
other instantiations since what is needed in a syntax check only.

The mentioned method is used in the "mcssvalidator" tool which outputs
information about the verified CSS files and exits with a code
according to the existence of problems (1) or not (0).