Fixes: NB#167434 - MThemeDaemon uses excessive amount of heap
RevBy: Armin Berres

Details:
The purpose of MThemeResourceManager has been to improve the performance
of SVG-loading by reusing existing SVG renderer instances. But keeping
the SVG renderer instances for some time leads to a very dynamic
allocation of the memory and some fragmentation. So even after all SVG
renderer instances have been deleted, the dirty memory of the themedaemon
is very high. As the SVGs get cached anyhow in the meantime, this kind of
performance optimization can be dropped now in favor of the memory demand.