Changes: Fix minimum size-hint of MLabel
RevBy: John Tapsell

Details: When using a simple text, the minimum size-hint of MLabel provided a wrong height. Like when using a rich text, the size constraint must get ignored and the minimum height should represent the height of one line. This patch fixes this issue and assures a consistent minimum size-hint no matter whether a rich text or a simple text is used.