Changes: Fix failing tests for ut_mlabel inside scratchbox
RevBy: TrustMe
Details: No failures for ut_mlabel are reported on the the
device, but several tests in Ut_MLabel::testSizeHint() fail
inside the scratchbox. The root cause is that MLabel uses
QStaticText internally, which does a custom layouting for
simple texts. Dependent on the used font this layout is
only "nearly equal" in comparison to a layout with QTextDocument.
Because of the different font metrics in the scratchbox
the result differs by 1 or 2 pixels. The unit test has been
fixed to use a similar layout mechanism to verify the result.