Changes: remove unused code from MLocale::setDefault()
RevBy: TrustMe

Details:

There was some code left in MLocale::setDefault() which was commented
out because I was unsure whether it is useful or not.

If enabled it would make MLocale::setDefault() change the gconf
keys for the locale and thus make MLocale::setDefault() have
a permanent effect and an effect on other applications.
I am pretty sure now that this does not make sense, i.e.
it should be removed for clarity.