Fixes: NB#194018 -  MPannableViewport::setMinimumSize resets panned widget position.
RevBy: Sergiy
Details: The panned widget inside pannable viewport is put into a layout.
When this layout tries to move and resize the panned widget it also resets
the position set by the pannable viewport. We connect a geometryChanged()
signal from the panned widget and if such cases occur, we set the position
to the correct one.

The same slot can be used with input panel widget relocation functionality,
MResizeListener is removed with this patch and the relocation code moved
to the slot inside MPannableViewportPrivate.