Changes: fix MLocale::countryEndonym() for some corner cases
RevBy: Berthold Krevert

Details:

http://site.icu-project.org/design/resbund/issues

causes problems for some country endonyms because a file like

    icu/source/data/region/sv_FI.txt

which contains only overrides for some countries

    sv_FI{
        Countries{
            UM{"USA:s yttre öar"}
        }
        Version{"2.0.45.99"}
    }

breaks the calculation of the country endonym for the countries not
contained in the override file, i.e.  the endonym for "FI" in Swedisch
is not found unless the file sv_FI.txt contains the complete list of
endonyms or the file sv.txt is tried as well.