Fixes: NB#195670 -  Calling show() on a MWindow that is already visible unsets the _NET_WM_STATE_SKIP_TASKBAR property
RevBy: Dominik

Details:
MTF sets and unsets _NET_WM_STATE_SKIP_TASKBAR property to hide windows from
switcher during some situations. But it does not take in account if the
property is also set/unst by external entities.

This patch makes MTF to take in account this situation, so when is going to
unset _NET_WM_STATE_SKIP_TASKBAR, but it was set by an external entity it
leaves it.