Fixes: NB#195924 - MVideoWidget does not work fullscreen with opengl or meego graphicssystem
RevBy: Michael Dominic K., Stanislav Ionascu, Daniel d'Andrada, Tomas Junnonen

Details:
When running with the meego graphicssystem the viewport itself is
no native native window. We use effectiveWinId() to grab the id
of the next parent window which is native.
Calling winId() on the viewport itself would make it a native window
which breaks the meego graphicssystem.