New: Floating list index similar to tooltip.
RevBy: Sergiy

Details: The floating list index will show a hovering tooltip with a hint,
which of the indexes is currently selected. The benefit of such a floating
list index, that it all list indexes fit into this helper. A special
parameter should be passed to MList::setIndexVisible(MList::Floating);

The number of hints is stylable, but a number of 3 is preferred.