Fixes: NB#195945 - Wrong label size in MDetailedListItem
RevBy: Armin Berres

Details: The size-hint for the preferred size returns an incorrect result if the current painting-rectangle is too small. In this case the text gets elided and a size-hint for the elided text is returned. The patch fixes the issue by always using the unelided version of the text do determine the preferred size-hint.