Changes: use @ as well as a delimiter when searching for translations
RevBy: Berthold Krevert

Details:

Without using @ as well as a delimiter, for the locale
“de@collation=phonebook” only de@collation=phonebook{,.qm} would be
found but if this does not exist de{,.qm} would not be tried as a
fallback.

Similar for “zh_CN@collation=pinyin”: if
zh_CN@collation=pinyin{,.qm} is not found, the next fallback should be
zh_CN{,.qm}. Without using @ as a delimiter the next fallback would be
zh{,.qm} already.