Fixes: NB#189611 - Incorrect Layout for MContentItem with small icon and small text
RevBy: Sergiy, Stanislav Ionascu

Details: Applying the size-hint QSizePolicy::MinimumExpanding for texts is not necessary to assure a non-clipped text. The layout of MContentItemView takes already care to respect the preferred size of the text-label. However applying the QSizePolicy::MinimumExpanding size-hint leads to a left-alignment if the the label has a restricted maximum size-hint (like in the testcase for NB#189611, where a style with restricted maximum size gets applied).