Fixes: NB#198551, Catalog installation and setting default locale gets slower by every installation
RevBy: Holger Schröder, Berthold Krevert

Details:

Make sure that previous installations of a catalog are removed first
in installTrCatalog() before trying to install a new catalog.  There is
no need to install the same catalog more then once with different
priorities.  One could skip the installation altogether if the catalog
is already installed, but it is better to remove the first instance,
then the priorities make more sense.