Changes: Fix crash when changing the theme
RevBy: Armin Berres

Details: Since 7a3847fe6f480962462d88c36dd796bcbf94c967 the MTheme
pixmap functions have been changed from asynchronous to synchronous.
This triggered a bug in MRemoteThemeDaemon when changing the theme
and might lead to a crash. The root cause of the crash is that
emitting the themeChangedSignal resulted in nested calls of
MRemoteThemeDaemonPrivate::waitForPacket(). This has been bypassed
with this patch by asynchronously emitting the signal.