Fixes: NB#198565 - MSeekBar is not signalling correctly press, value changed and release
RevBy: Tomas Junnonen

Details:
When clicking (quick press and release) on the slider, the animation was updating its value.
This would make the widget to give a value different than what was set from the click if
its queried for example in the mouse release signal. This also changes a bit the concept of
the widget's value because the value indicated in fact the position of the slider at a given
moment. Summarizing, the animation shouldn't control the value.

These changes introduce a "position" property which controls the widget's position and may be
controlled by the animation. The position should always be equal to the value when no animation
is set.