Fixes: NB#199168 - Multilength strings trigger loading of all the installed fonts -> huge memory usage from nokia/arabic ttf fonts
RevBy: Thomas Mönicke, Armin Berres, Stanislav Ionascu

Details: If a string containing characters that are not part of the used font is passed to QFontMetrics, all the other available fonts are loaded until the glyph is found. In the case of the multilength-length character represented by 0x9c no glyph is available in any font, so all fonts are loaded. The patch fixes this issue by passing only the first (= longest by convention) string to QFontMetrics.