Changes: fix and enable focus-related MTextEdit tests
RevBy: Viacheslav Sobolev, Michael Hasselmann
Details: MTextEditPrivate::requestSip() insists on QApplication::focusWidget()
returning a widget or it bails out.  For QApplication::focusWidget() to
return a widget the system relies on XFocusIn event or the focused widget
being set explicitly.  The event never arrives if mcompositor is not running
and might also arrive too late for the purposes of the unit tests even if it
is running, so we'll set the widget explicitly now.