Fixes: NB#193775 - White flash is shown for a while, when a background application is invoked from content manager
RevBy: Armin Berres

Details: If an application receives the signal MWindow::switcherExited() and is not able to redraw the content within a very short time, a white background is drawn temporary. This is fixed by setting the widget attribute Qt::WA_NoSystemBackground (the already set attribute Qt::WA_OpaquePaintEvent is not sufficient for this usecase).