Changes: try harder to load most used pixmaps
RevBy: Holger Schröder, Stanislav Ionascu

Details:
Loading of the most used pixmaps should be finished as fast as possible
while not rendering the device unusable at the same time.
We not sleep 100ms between loading two images and make sure the device
is 70% idle.
As soon as we start e.g. scrolling in an app we suspend loading.