Fixes: NB#159706 - Blocking of background application paints not working
RevBy: Stanislav Ionascu

Details:
We will not discard all paint events if MWindow::isOnDisplay() returns
false. If the visibility information is wrong for whatever reason this
might lead to a visible app not rendering anything. Close attention must
be paid therefor that no one breaks this.
One thing should be noted: Applications should not paint in the first
place when they are not visible. This change is just for those not
behaving nicely.