Fixes: NB#181399 - Incremental layouting during startup
RevBy: Stanislav Ionascu, Sergiy Dubovik

Details: Originally the call to updateListGeometry() has been postponed to the next event loop, as otherwise too many cells get created without having a pannable viewport. But this results in showing a blank content when showing the list initially. The patch checks whether a pannable viewport is already available and updates the list geometry synchronously in this case.