Changes: allow some paint events when being invisible
RevBy: Stanislav Ionascu

Details:
Even when the app is invisible we sometimes receive paint events e.g.
right before we are getting visible again. At this point in time the
visibility information might not have reached us yet.
Apart from that by allowing a small number of paint events the whole
thing becomes a bit more robust.
When an app needs more than the allowed 5 paint events when it is
invisible something is really going wrong and should be fixed.