Changes: replace strings by ids in style classes and style attributes
RevBy: Armin Berres, Peter Penz

Details:
Strings for names, types and keywords are passed around as unique integers
to reduce overhead with string operations.
The strings and the mapping to ids are stored in a mmaped file. The memory
of all strings will be shared between application. This part has been
written by Armin Berres.