Fixes: NB#195051 - Destructor causes crash in buttons
RevBy: Stanislav Ionascu, Michael Hasselmann
Details:
During the destruction of a widget (in the case of this specific bug, a MContainer)
that contains several buttons in a button group, when the checked button gets
deleted before the others and the group itself, the group will automatically check
one of the remaining buttons. Doing this will end up triggering the theming system
which will crash because it's querying the button's parent which is in the middle
of its destruction.

This patch solves this by deferring the exclusivity validation of the button group
to a subsequent event in the event loop.