Fixes: NB#194018 - MPannableViewport::setMinimumSize resets panned widget position.
RevBy: Dominik
Details: setting position of the panned widget was working around pannable
viewport layout. When the layout wanted to update geometry of the panned widget
it was resetting paned widget position. This patch removes the workaround in
the pannable viewport and moves handling of the position change to the
pannable viewport layout.