Fixes: NB#203504 - Not possible to disable orientation change animation in MWindow
RevBy: Stanislav, Dominik
Details:
Added a new dynamic property to MWindow that enables an application
to choose whether it wants the orientation change transitions generated by
a MWindow to be animated or not.
It's made as a dynamic property (instead of a hard-coded one) since it's
only needed for a specific application to achieve a so far unique use case.