Fixes: NB#204735 - Emulation of pinch gesture not working
RevBy: Dominik
Details: because of a change in Pinch gesture recognizer in Qt,
we now have to send events to the current QGraphicsView, not to the scene.
Also, we need to create gesture context inside QGraphicsView and not
on the viewport of it. The "testFlag()" use in mscene will remove
the bug that sometimes the emulation wouldn't work until pressing alt key.