Changes: draw with 50% transparency when MButtonIcon is clicked
RevBy: Stanislav Ionascu

Details:
The current glow and shrink effect is not specified anywhare. Apart
from that the implementation is inefficient.
Now the icon will be drawn with transparency when the icon button
is clicked.

Changes: Cleanup of button icon view classes, support for label to get the content opacity.

RevBy: Armin Berres