Changes: cache QVariants globally
RevBy: TrustMe

Details:
The last four commits were not meant to be committed at this very
moment. I still wanted to make the change which is in this commit now.
Nevertheless they were reviewed by Peter Penz.
This commit introduces a global QVariant cache for MStyleSheetAttribute.
This reduces the number of created QVariants at startup by a factor
of 10. At runtime nearly no new QVariants need to be created now.