Fixes: NB#198347 - Remove Animated layout demos
RevBy: Mike FABIAN

Details:

I wrote this ages ago to give myself a testbed for the layouts.  It
has bitrotted greatly since then, and causing compile failures when
compiling with minimal options.  It doesn't fit in natively on the
device, the images use the old libdui logo, and so on.

Since it's so broken and noone has ever complained about it, I'm just
deleting it.