Fixes: NB#206036 - When minimized combobox shows some additional lines.
RevBy: Stanislav Ionascu

Details: We are now switching rendering modes in runtime, minimized applications
are drawn in software mode.

MScalableImage was using MComponentDate::softwareRendering to determine
if software mode is enabled. But this method checks only if -software was given
in command line (or in environment variable), not if current rendering mode
is software.

Now paintEngine()->type() is used to determine which rendering mode
is currently active.