Fixes: NB#207292 - Unable to resize MDialog Buttons using styling
RevBy: Dominik
Details: Button width for MDialog was changing value in the way that
buttons were too small to draw, because of expanding spacer widgets on
both sides of button box. Currently button box removes spacers if button
width is set to -1 (for auto expanding to full width of button box) even
if button box centering is on. Additionally if style property is set
differently for Landscape and Portrait modes both versions are respected
now, not only the one from orientation in which application was started.