Fixes: NB#207150, Invalid entries are ignored when the pasted value starts with valid number
RevBy: Berthold Krevert

Details:

Make parsing of numbers stricter. It is still not as strict
as the number parsing of QLocale, but much closer now.
Input which has trailing garbage is considered as invalid now,
i.e. when the input cannot be parsed until the end this is an error
For example parsing

   4f

gives an error now instead of the result "4".

This also means that when trying to parse integers, a decimal
separator will cause an error, for example parsing

   1.5

as an integer will cause a parse error instead of returning 1.