Changes: Allow MWidgetStyle definitions to be used outside "core" and "application" styles.
RevBy: Sergiy Dubovik, Armin Berres

Details: The MWidgetController and MStylableWidget, both use the MWidgetStyle base class,
which is defined in "core" library, thus styles that are defined outside the "core" and
"application" CSSes cannot be applied. Including the Common styles.
The fix esures that for the base "MWidgetStyle" class, the style definitions from all libraries
are used.