Fixes: NB#205533
RevBy: Maciej Jabłoński, Dominik Kapusta
Details: System modal dialog creates its own scenemanager. When it contains combobox and
combobox was clicked showing popuplist, popuplist is displayed using system modal dialog's
sceneManager. When system modal dialog is destroyed firstly MWindow was destroyed,
(scenemanager with it) which caused popuplist also be deleted. Then dialog itself was destroyed
and so it's contents including combobox, which tried to also delete popuplist which belongs to
MComboBoxView. So we had crash here.

This issue was fixed in two places: in mdialog and in mcomboboxview. Each of this fix alone
is good enough for this particular case, but fixing it in two places may also cover
some crazy future use cases.
In mdialog i changed order of deletion, dialog itself is deleted first and then MWindow
(it is done via deleteLater). So MComboBox will be destroyed first and will delete popuplist
which will be removed from scene - during deleting of mwindow it won't be deleted again.

In MComboBoxViewPrivate i used QPointer instead of oridinary pointer to hold MPopupList.
This way if it was deleted before somewhere else (read: during deletion of scene),
QPointer::data will be set to NULL and deleting it again will be no op.