Fixes: NB#208549 - ft_mmoc aborts
RevBy: Mike Fabian
Details:
For some reason the qt moc is not installed on the
device in the current test environment. Without it
the ft_mmoc tests cannot be executed. This patch
skips this test in this case. This should fix this
bug. Also this test does not need to link against
QtGui, so i changed it to use QCoreApplication
for the appliation directory path.