Changes: temporarily disable blocking of paint events with meego graphicssystem
RevBy: Stanislav Ionascu

Details:
Right now blocking of applications running with the meego graphicssystem
causes the meego graphicssystem to clear the buffer with black.
As long as we have a solution for this we disable blocking the paint events.

This only affects fullscreen applications, minimized ones are using the
raster graphicssystem.