Fixes: MTextEdit: pre-edit injection event passing in the presence of QMultiInputContext
RevBy: Pekka Vuorela
Details: after disabling explicit loading of MInputContext and relying on
QT_IM_MODULE, the actual input context returned by QApplication::inputContext
is QMultiInputContext which acts as a proxy and forwards calls to, in our
typical case, MInputContext.  But it doesn't override and forward
QObject::event(), so we can't use QCoreApplication::sendEvent for sending the
pre-edit injection event to the input context.