Fixes: NB#210072 - Widgetsgallery grid page should stop loading images when page is left
RevBy: Armin Berres

Details: Pauses the image loading during list paning and when leaving page, resumes as soon
as the page appears, or the list panning is stopped.
Tries to load first the visible items of the list, then starts from the end of the queue. In case
if we're in the begining of the list, starts from the top.