Changes: allow nothing, pixmap or color as default background filling at app startup
RevBy: Peter Penz, Stanislav Ionascu

Details:
So far when running with the meego graphicssystem we fill all windows
with a pixmap specified via the theming system. If no pixmap is found
the window will be initialized with black.
With this patch we do not fill with black but with the actual color set
via the style.

If an application does not want the background filling it can disable it
with a CSS snippet like this:
/----
MWindowStyle {
    x11-startup-pixmap: ;
    startup-fill-color: ;
}

MWindowStyle.Landscape {
    x11-startup-pixmap: ;
    startup-fill-color: ;
}
\----
This is interesting e.g. for all applications starting during boot to keep
the framebuffer intact.