Fixes: NB#210881, calendar app crashing on trying to add new event
RevBy: Holger Schröder

Details:

Fix it by using

 mutable QCache<QString, icu::DateFormat> _dateFormatCache;

instead of a pointer as the member variable in MLocalePrivate.

Adding a

   _dateFormatCache = 0;

after the

   delete _dateFormatCache;

in the operator= and adding a “delete _dateFormatCache;” in the
destructor would have fixed the problem as well of course, but not
using a pointer also fixes the problem.

My idea of using the pointer was to avoid calling the constructor of
QCache unless really needed to make construction of MLocale objects
cheaper. Maybe a MLocale object doesn’t need date and time
formatting at all, then it also doesn’t need that cache.

But Holger convinced me that creating a QCache is very cheap.