Fixes: NB#210713, Text Input Field is "jumping"
RevBy:  Michael Hasselmann, Michał Guminiak
Details: If panning position is out of viewport because of viewport
resize we should not animate but rather set position directly to
a valid one. This was already done but now with range height extension
in use we have to take it into account, too.