Fixes: NB#207188 - Coreweb: random search crash in QPixmap::pixmapData
RevBy: Stanislav Ionascu

Details:
The original root cause of the crash might be already fixed by f33401d7017177314aeb322ca6052506d012bc0a.
But the values might be NULL, in case if the style was not initialized properly.
So ensure that pixmaps are not NULL before trying to draw them.