Fixes: NB#211689 - Unable to switch style in MBasicListItem at runtime
RevBy: Stanislav Ionascu

Details: Enables the capability to switch from different basic list item styles
in runtime without setting again the properties.

Merged with changes by: Stanislav Ionascu

Details: On style change the items are going to be removed from layout and scene.
As soon as the item is added into the layout it will be added into the scene again.
Private of the basic list item will take care of deleting the child items.