Changes: Add slider/seeker style properties.
RevBy: Stanislav Ionascu

Details:
Partially based on work by Daniel Borgmann.

This adds the style properties elapsed-offset, groove-margin,
indicator-offset, and indicator-stylename which are required for
accurate styling of the slider and seeker.

This is also required for the style updates in NB#209116.

ModifiedBy: Stanislav Ionascu
Details:
Slightly refactored the name of the setStyleName method in SliderIndicator widget,
for more accurate description.