Fixes: NB#215046 - Regression: MeeGoTouch eats touchpoints
RevBy: Marcin, Dominik
Details: When user presses the touch screen with two or more fingers,
we are going to stop the initial press timer and send events to the
scene. This will ensure that the scene will send the information about
multiple fingers touch event to the widgets in it.