Changes: fix directory pathes for windows
RevBy: TrustMe

Details: under windows path names like
//var/cache... are interpreted as directory
"cache" on file server "var". and when
we try to access these, windows tries
to connect to the server, and most times
this does not work and the connect will
timeout, for me in around 20 seconds.

by not accessing non-existing remote servers
the widgetsgallery startup is happening
in less than a second again.