Changes: MScalableImage: Remove cached pixmap from QPixmapCache manually
RevBy: Armin Berres

Details: Currently the scalable image gets cached in QPixmapCache if it is not rendered by an OpenGL-backend. But it does not get removed manually, which leads to an unnecessary increasing of QPixmapCache. Although QPixmapCache internally takes care to release pixmaps that are not used during a specific period of time, not manually releasing results in several MB of additional peak for some usecases.