Fixes: NB#208270 - Status area is displayed in portrait orientation when application is in landscape and vice-verse.
RevBy: Dominik
Details: Now when handling current app window change we are testing for its orientation
(_MEEGOTOUCH_ORIENTATION_ANGLE) right away instead of waiting for PropertyChangeEvent.
This is necessary because non-lmt windows will not set this property, and app will not get updated.

Added unit test for case when non-lmt app is on top.