Changes: Make windows follow the rotation of a remote screen (when present)
RevBy: Daniel d'Andrada

ModifiedBy: Daniel d'Andrada
 - surrounded MOrientationTrackerPrivate::rotateWindows() with HAVE_CONTEXTSUBSCRIBER
 - removed unnecessary MOrientationTrackerPrivate::remoteTopEdgeInfo

Details:
RemoteScreen.TopEdge context property is defined as follows:
"""
The location of the top edge of the remote screen relative to the main one.
Possible values are the strings "top", "bottom", "left", and "right", which
refer to the edges of the main display screen in its normal orientation as
defined by the graphics hardware.
"""

When a remote screen is present, our windows will follow its rotation (if possible).
I.e. the rotation of a remote screen (relative to our screen) dictates the
rotation of our windows.

This patch doesn't require the new property to be available in which case the
functionality is unchanged.