Fixes: NB#209179 - QDBusConnection created before MApplication instantiation leads to threading issues
RevBy: Armin Berres

Details: According to the qt bugtracker, the behaviour of
QDBusConnection::sessionBus() and QDBusConnection::systemBus()
is undefined, when they are used before the QApplication
construnctor was executed.

Because of this we defer the creation of the sessionBus
until it is accessed the first time. And we do write an
error message when the sessionBus is accessed before
the qApp instance exists.