Changes: add benchmark pt_mcharsetdetector
RevBy: Holger Schröder, Armin Berres

Details:

Whether hints for the encoding or the locale are given and whether
input filtering is enabled seems to be insignificant.

The time needed for the detection increases with the length
of the input string (longer input gives of course a more
reliable detection). On scratchbox x86 (Lenovo T500), with
Chinese input the time for detection is approximately:

    4 characters:  0.01 msecs
   40 characters:  0.02 msecs
  400 characters:  0.1  msecs
 4000 characters:  1    msecs
40000 characters: 10    msecs