Changes: Make scene window opacity independent from its layer effect
RevBy: Stanislav Ionascu
Details: Since the layer effect must set its opacity independently
from the corresponding scene window, we must have them as siblings
instead of having the layer as the parent of the scene window.

That's different from making the layer effect not propagate its
opacity to the scene window child. When we change the opacity of the
root element during an orientation change animation, we want both
the layer effect *and* its scene window to obey the set opacity,
whereas during a dialog appearance animation we want the layer opacity
to change independently from the dialog itself. Without this change
setting the opacity of the root element to zero will have no effect
over an open dialog.

That will improve the appearance of the rotation of system dialogs.
This is the last commit needed to completely fix NB#218983
(Rotation with open dialog is ugly).