Fixes: NB#221617 - Each MAction adds approximately 5ms to every MApplicationPage appearance (Using gallery css)
RevBy: Armin Aberres, Jukka Halonen, Adrian Yanes.

Details:
Every time an action was added, a new button had to being created; these
buttons were being deleted after the actions were removed. Since the
actions are removed when a page is hidden, this meant that hidding and
showing a page would always need to create new widget, thus taking time.

These changes add the actions' buttons to a QCache instead of deleting them.
This way, when an action is added, first it checks if the corresponding
button is already created which saves time.