Fixes: NB#221613 - Minimal Mapplication is repainted 21-24 times on startup
RevBy: Armin Berres

Details: If a position-indicator is shown after starting an application, it will get faded out which results in around 20 repaint operations. In case where the content is fully visible the painting code won't draw anything. The patch fixes the issue that the animation will get started in this case, as the indicator is anyhow invisible.