Fixes: NB#218196 - Not enough mouse move events and too big distance between the mouse press and first mouse move. Fixes: NB#216904 - MScene delays touch events for every widget Fixes: NB#185681 - Too much time is spent between click on booster-enabled application icon and start of application's main function.
RevBy: Stanislav Ionascu, Daniel d'Andrada
Details: The touch and mouse event delay in mscene is removed and a state machine
to handle highlighting in mlistitem is introduced.

Also this patch fixes a regression in switch button if these delays are removed:
The tap on the switch button wasn't causing state change.

When the user taps quickly on a switch button, the mouse move
events can be generated and switch button view was automatically switching
into "drag" mode.

It was working alright before because the scene was delaying events and
if the user only tapped on the button, then the move events weren't delivered
to the button.