Changes: Improve performance for orientation changes
RevBy: John Tapsell

Details: QGraphicsLayout::activate() causes all items
in the layout to be immediately rearranged. This is
an unnecessary operation in case of an orientation
change not all items in the layout have already updated
their style. The performance gets improved noticable
when just invalidating the layout and doing the
rearranging of the items in one step when receiving the
LayoutRequest event later.