Fixes: NB#225559 - Timezone applet leads to settings crash
RevBy: Thomas Mönicke
MergeRequest: 1375

Details: When the content is filtered, persistent indexes are changed, the model emits non-animated rows(Removed|Inserted), and list tries to access no longer valid index to detect animation range.
The fix ensures early out if the request was non-animated or animation is not supported by the type of the data model.